serializable module file renamed
