+++ example/app/bootstrap/bootstrap-example.component.html
@@ -22,7 +22,7 @@
 
     </form>
 
-    <button (click)="test()" type="button">Test</button>
+    <!--<button (click)="test()" type="button">Test</button>-->
 
 </div>
 <!--
