+++ modules/core/src/model/dynamic-form-value-control.model.ts
@@ -1,7 +1,7 @@
 import {ValidatorFn, AsyncValidatorFn} from "@angular/forms";
 import {Subject} from "rxjs/Subject";
 import {DynamicFormControlModel, DynamicFormControlModelConfig, ClsConfig} from "./dynamic-form-control.model";
-import {serializable} from "../decorator/serialize.decorator";
+import {serializable} from "../decorator/serializable.decorator";
 import {getValue, isDefined} from "../utils";
 
 export interface DynamicFormValueControlModelConfig extends DynamicFormControlModelConfig {
