+++ modules/core/src/model/input/dynamic-input.model.ts
@@ -1,6 +1,6 @@
 import {ClsConfig} from "../dynamic-form-control.model";
 import {DynamicInputControlModel, DynamicInputControlModelConfig} from "../dynamic-input-control.model";
-import {serializable} from "../../decorator/serialize.decorator";
+import {serializable} from "../../decorator/serializable.decorator";
 import {getValue} from "../../utils";
 
 export const DYNAMIC_FORM_CONTROL_TYPE_INPUT = "INPUT";
