+++ modules/core/src/model/radio/dynamic-radio-group.model.ts
@@ -1,7 +1,7 @@
 import {ClsConfig} from "../dynamic-form-control.model";
 import {DynamicOptionControlModel, DynamicOptionControlModelConfig} from "../dynamic-option-control.model";
 import {DynamicFieldSet} from "../form-group/dynamic-form-group.model";
-import {serializable} from "../../decorator/serialize.decorator";
+import {serializable} from "../../decorator/serializable.decorator";
 import {getValue} from "../../utils";
 
 export const DYNAMIC_FORM_CONTROL_TYPE_RADIO_GROUP = "RADIO_GROUP";
